refactor(decompose): split webapp/e2e-tests/electron/critical_e2e_verification_tests/electron-project-selection.spec.ts into folder
