+++ webapp/e2e-tests/electron/critical_e2e_verification_tests/electron-project-selection/paths.ts
@@ -1,0 +1,6 @@
+import * as path from 'path';
+
+export const PROJECT_ROOT = path.resolve(process.cwd());
+export const CI_FLAGS = process.env.CI
+    ? ['--no-sandbox', '--disable-dev-shm-usage', '--use-gl=angle', '--use-angle=swiftshader']
+    : [];
