+++ webapp/e2e-tests/electron/critical_e2e_verification_tests/electron-project-selection/types.ts
@@ -1,0 +1,7 @@
+import type { Core as CytoscapeCore } from 'cytoscape';
+import type { ElectronAPI } from '@/shell/electron';
+
+export interface ExtendedWindow {
+    cytoscapeInstance?: CytoscapeCore;
+    electronAPI?: ElectronAPI;
+}
